part 4 completed

+++ part4/blog-list-app/controllers/blogs.js
@@ -1,12 +1,19 @@
 const blogsRouter = require("express").Router();
 const Blog = require("../models/blog");
+const User = require("../models/user");
+const jwt = require("jsonwebtoken");
+const middleware = require("../utils/middleware");
 
 // GET methods
 blogsRouter.get("/", async (request, response) => {
   //Blog.find({}).then((blogs) => {
   //  response.json(blogs);
   //});
-  const blogs = await Blog.find({});
+  const blogs = await Blog.find({}).populate("user", {
+    username: 1,
+    name: 1,
+    id: 1,
+  });
   response.json(blogs);
   response.status(500).end();
 });
@@ -14,6 +21,10 @@ blogsRouter.get("/", async (request, response) => {
 // POST methods
 blogsRouter.post("/", async (request, response) => {
   const body = request.body;
+  const user = request.user;
+  if (!request.user) {
+    return response.status(401).json({ error: "token missing or invalid" });
+  }
 
   if (!body.title || !body.url) {
     return response.status(400).json({ error: "Title and URL are required" });
@@ -24,15 +35,35 @@ blogsRouter.post("/", async (request, response) => {
     author: body.author,
     url: body.url,
     likes: body.likes || 0,
+    user: user.id,
   });
+
   const savedBlog = await blog.save();
+  user.blogs = user.blogs.concat(savedBlog._id);
+  await user.save();
   response.status(201).json(savedBlog);
 });
 
 // DELETE methods
 blogsRouter.delete("/:id", async (request, response, next) => {
-  await Blog.findByIdAndDelete(request.params.id);
-  response.status(204).end();
+  try {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET);
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: "token invalid" });
+    }
+    const blog = await Blog.findById(request.params.id);
+    if (!blog) {
+      return response.status(404).json({ error: "blog not found" });
+    }
+
+    if (blog.user.toString() !== decodedToken.id) {
+      return response.status(401).json({ error: "unauthorized user" });
+    }
+    await Blog.findByIdAndDelete(request.params.id);
+    response.status(204).end();
+  } catch (error) {
+    next(error);
+  }
 });
 
 // PUT methods
